+++ NotesServer/pom.xml
@@ -32,13 +32,16 @@
 			<artifactId>flyway-core</artifactId>
 			<version>7.8.2</version>
 		</dependency>
+		<!-- https://mvnrepository.com/artifact/org.flywaydb/flyway-mysql -->
 		<dependency>
 			<groupId>org.flywaydb</groupId>
 			<artifactId>flyway-mysql</artifactId>
+			<version>9.19.4</version>
 		</dependency>
+
 		<dependency>
 			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-web</artifactId>
+m			<artifactId>spring-boot-starter-web</artifactId>
 		</dependency>
 
 		<dependency>
