+++ NotesServer/src/main/resources/application.properties
@@ -1,14 +1,9 @@
 jwt.secret=secret
-#spring.datasource.url=jdbc:mysql://localhost/notesMonesh?createDatabaseIfNotExist=true&autoReconnect=true&useSSL=false
-#spring.datasource.username=root
-#spring.datasource.password=root
-#spring.datasource.platform=mysql
-#spring.jpa.hibernate.ddl-auto=update
 spring.datasource.driverClassName = com.mysql.jdbc.Driver
 spring.datasource.url = jdbc:mysql://localhost:3306/notes_todo?autoreconnect=true
 spring.datasource.username = root
 spring.datasource.password = root
 flyway.url = jdbc:mysql://localhost:3306/mysql
-flyway.schemas = flyway_test
+flyway.schemas = notes_todo
 flyway.user = root
 flyway.password = root
