+++ notesuireact/src/models/authlogin.ts
@@ -2,7 +2,7 @@ import { deserialize, serializable,serialize } from "serializr";
 export class Authlogin{
 
   @serializable
-  public emailId : string ="";
+  public username : string ="";
   @serializable
   public password : string ="";
   deserialize(input: any): this {
